server/requirements.txt
django==4.1
djangorestframework==3.13.1
psycopg2==2.9.3
drf-yasg==1.21.3
django-cors-headers==3.13.0